chore: move SKCanvasVisual to a separate partial
